+++ src/groups/killmails.rs
@@ -15,43 +15,43 @@ pub struct RecentKillMail {
 #[derive(Debug, Deserialize)]
 #[allow(missing_docs)]
 pub struct KillmailAttacker {
-    pub alliance_id: Option<u64>,
-    pub character_id: Option<u64>,
-    pub corporation_id: Option<u64>,
-    pub damage_done: u64,
+    pub alliance_id: Option<i32>,
+    pub character_id: Option<i32>,
+    pub corporation_id: Option<i32>,
+    pub damage_done: i32,
     pub final_blow: bool,
     pub security_status: f64,
-    pub ship_type_id: Option<u64>,
-    pub weapon_type_id: Option<u64>,
+    pub ship_type_id: Option<i32>,
+    pub weapon_type_id: Option<i32>,
 }
 
 #[derive(Debug, Deserialize)]
 #[allow(missing_docs)]
 pub struct KillmailItem {
-    pub flag: u64,
-    pub item_type_id: u64,
-    pub quantity_destroyed: Option<u64>,
-    pub quantity_dropped: Option<u64>,
-    pub singleton: u32,
+    pub flag: i32,
+    pub item_type_id: i32,
+    pub quantity_destroyed: Option<i64>,
+    pub quantity_dropped: Option<i64>,
+    pub singleton: i32,
 }
 
 #[derive(Debug, Deserialize)]
 #[allow(missing_docs)]
 pub struct KillmailVictim {
-    pub alliance_id: Option<u32>,
-    pub character_id: Option<u64>,
-    pub corporation_id: Option<u64>,
-    pub damage_taken: u64,
-    pub faction_id: Option<u64>,
+    pub alliance_id: Option<i32>,
+    pub character_id: Option<i32>,
+    pub corporation_id: Option<i32>,
+    pub damage_taken: i32,
+    pub faction_id: Option<i32>,
     pub items: Option<Vec<KillmailItem>>,
 }
 
 #[derive(Debug, Deserialize)]
 #[allow(missing_docs)]
 pub struct Killmail {
-    pub killmail_id: u64,
+    pub killmail_id: i32,
     pub killmail_type: String,
-    pub solar_system_id: u64,
+    pub solar_system_id: i32,
     pub moon_id: Option<i32>,
     pub attackers: Vec<KillmailAttacker>,
     pub victim: KillmailVictim,
@@ -65,7 +65,7 @@ impl<'a> KillmailsGroup<'a> {
         "get_characters_character_id_killmails_recent",
         RequestType::Authenticated,
         Vec<RecentKillMail>,
-        (character_id: u64) => "{character_id}"
+        (character_id: i32) => "{character_id}"
     );
 
     api_get!(
@@ -74,7 +74,7 @@ impl<'a> KillmailsGroup<'a> {
         "get_killmails_killmail_id_killmail_hash",
         RequestType::Public,
         Killmail,
-        (killmail_id: u64) => "{killmail_id}",
+        (killmail_id: i32) => "{killmail_id}",
         (killmail_hash: &str) => "{killmail_hash}"
     );
